Modify Record signing to use a passed in `sign` function
