+++ test/modules/multipart/mpsingle.test.js
@@ -1,6 +1,8 @@
+import {sign} from "bitcoinjs-message"
 import bitcoin from 'bitcoinjs-lib'
 import {flo_testnet} from '../../../src/config/networks'
 import MPSingle from '../../../src/modules/multipart/mpsingle'
+import ExplorerWallet from '../../../src/modules/wallets/ExplorerWallet'
 import Oipindex from '../../../src/core/oipd-api/daemonApi'
 
 const index = new Oipindex()
@@ -73,12 +75,30 @@ describe("MPSingle", () => {
 		})
 	})
 	describe('Signing', () => {
-		it('Create and validate signature', () => {
+		it('Create and verify signature', async () => {
 			let network = flo_testnet.network
 			let ECPair = bitcoin.ECPair.makeRandom({network})
+
+			let signMessage = async (message) => {
+				let privateKeyBuffer = ECPair.privateKey;
+
+				let compressed = ECPair.compressed || true;
+
+				let signature_buffer
+				try {
+					signature_buffer = sign(message, privateKeyBuffer, compressed, ECPair.network.messagePrefix)
+				} catch (e) {
+					throw new Error(e)
+				}
+
+				let signature = signature_buffer.toString('base64')
+
+				return signature
+			}
+			
 			let address = bitcoin.payments.p2pkh({pubkey: ECPair.publicKey, network}).address
 			let mps = new MPSingle({part: 0, max: 1, reference: 'reference', address, data: 'data'})
-			let {success, signature, error} = mps.signSelf(ECPair)
+			let { success, signature, error } = await mps.signSelf(signMessage)
 
 			expect(success).toBeTruthy()
 			expect(signature).toBeDefined()
